[UPDATE] Use redux hooks :fire:.
