+++ app.json
@@ -1,7 +1,7 @@
 {
   "expo": {
     "name": "Abacus",
-    "description": "Abacus: a FireflyIII iOS Client",
+    "description": "🔥 Abacus: Firefly III iOS app 🔥",
     "slug": "abacus",
     "privacy": "hidden",
     "platforms": [
@@ -16,7 +16,8 @@
     },
     "updates": {
       "enabled": true,
-      "checkAutomatically": "ON_ERROR_RECOVERY"
+      "checkAutomatically": "ON_ERROR_RECOVERY",
+      "url": "https://u.expo.dev/292ed6dc-804c-4444-95f5-fa5d76d9913b"
     },
     "ios": {
       "buildNumber": "0.2.1",
@@ -29,6 +30,9 @@
       }
     },
     "scheme": "abacusiosapp",
-    "githubUrl": "https://github.com/victorbalssa/abacus"
+    "githubUrl": "https://github.com/victorbalssa/abacus",
+    "runtimeVersion": {
+      "policy": "sdkVersion"
+    }
   }
 }
