+++ src/containers/Home.tsx
@@ -1,54 +1,26 @@
-import React, { useEffect } from 'react';
-import { connect } from 'react-redux';
+import React, { useEffect, FC } from 'react';
+import { useSelector, useDispatch } from 'react-redux';
 import { useToast } from 'native-base';
 
-import Layout from '../native/components/Home';
-import { HomeDisplayType } from '../models/firefly';
+import Layout from '../components/Home';
+import { RootDispatch, RootState } from '../store';
 
-const mapStateToProps = (state) => ({
-  loading: state.loading.models.firefly,
-  netWorth: state.firefly.netWorth,
-  spent: state.firefly.spent,
-  balance: state.firefly.balance,
-  earned: state.firefly.earned,
-});
-
-const mapDispatchToProps = (dispatch) => ({
-  handleChangeRange: dispatch.firefly.handleChangeRange,
-  getSummary: dispatch.firefly.getSummaryBasic,
-  getDashboard: dispatch.firefly.getDashboardBasic,
-});
-
-type HomeContainerType = {
-  loading: boolean,
-  netWorth: HomeDisplayType[],
-  spent: HomeDisplayType[],
-  balance: HomeDisplayType[],
-  earned: HomeDisplayType[],
-  getSummary: () => Promise<void>,
-  getDashboard: () => Promise<void>,
-  handleChangeRange: (value?: object) => Promise<void>,
-};
-
-const Home = ({
-  loading,
-  netWorth,
-  spent,
-  balance,
-  earned,
-  getSummary,
-  getDashboard,
-  handleChangeRange,
-}: HomeContainerType) => {
+const Home: FC = () => {
   const toast = useToast();
+  const loading = useSelector((state: RootState) => state.loading.models.firefly);
+  const firefly = useSelector((state: RootState) => state.firefly);
+  const dispatch = useDispatch<RootDispatch>();
 
   useEffect(() => {
-    (async () => handleChangeRange())();
+    dispatch.firefly.handleChangeRange({});
   }, []);
 
   const fetchData = async () => {
     try {
-      await Promise.all([getSummary(), getDashboard()]);
+      await Promise.all([
+        dispatch.firefly.getSummaryBasic(),
+        dispatch.firefly.getDashboardBasic(),
+      ]);
     } catch (e) {
       toast.show({
         placement: 'top',
@@ -61,13 +33,13 @@ const Home = ({
   return (
     <Layout
       loading={loading}
-      netWorth={netWorth}
-      spent={spent}
-      balance={balance}
-      earned={earned}
+      netWorth={firefly.netWorth}
+      spent={firefly.spent}
+      balance={firefly.balance}
+      earned={firefly.earned}
       fetchData={fetchData}
     />
   );
 };
 
-export default connect(mapStateToProps, mapDispatchToProps)(Home);
+export default Home;
