--- src/native/components/UI/RangeTitle.tsx
@@ -1,96 +1,0 @@
-import React from 'react';
-import {
-  HStack,
-  Box,
-  Text,
-  Select,
-  CheckIcon, IconButton,
-} from 'native-base';
-import { connect } from 'react-redux';
-import moment from 'moment';
-import { AntDesign } from '@expo/vector-icons';
-import * as Haptics from 'expo-haptics';
-import { Dispatch, RootState } from '../../../store';
-
-const RangeTitle = ({
-  start,
-  end,
-  range,
-  rangeTitle,
-  handleChangeRange,
-}) => (
-  <Box
-    shadow={2}
-    backgroundColor="white"
-    pl={3}
-    pr={3}
-    zIndex={30}
-    safeAreaTop
-  >
-    <HStack justifyContent="space-between" alignItems="center">
-      <IconButton
-        shadow={2}
-        borderRadius={15}
-        variant="solid"
-        _icon={{
-          as: AntDesign,
-          name: 'arrowleft',
-        }}
-        onPress={() => handleChangeRange({ direction: -1 })}
-        onTouchStart={() => Haptics.impactAsync(Haptics.ImpactFeedbackStyle.Light)}
-      />
-      <Box pl={2} mt={1} height={41} width={170} justifyContent="center">
-        <Text style={{ fontFamily: 'Montserrat_Bold', fontSize: 21 }}>
-          {rangeTitle}
-        </Text>
-        <Text style={{ fontSize: 12 }}>
-          {`${moment(start).format('ll')} - ${moment(end).format('ll')}`}
-        </Text>
-      </Box>
-      <Select
-        m={1}
-        borderRadius={15}
-        variant="outline"
-        dropdownIcon={<></>}
-        height={41}
-        width={84}
-        _selectedItem={{
-          bg: 'primary.600',
-          endIcon: <CheckIcon size="5" />,
-        }}
-        selectedValue={`${range}`}
-        onValueChange={(v) => handleChangeRange({ range: v })}
-      >
-        <Select.Item key="1" label="Monthly" value="1" />
-        <Select.Item key="3" label="Quarterly" value="3" />
-        <Select.Item key="6" label="Semiannually" value="6" />
-        <Select.Item key="12" label="Yearly" value="12" />
-      </Select>
-
-      <IconButton
-        shadow={2}
-        borderRadius={15}
-        variant="solid"
-        _icon={{
-          as: AntDesign,
-          name: 'arrowright',
-        }}
-        onPress={() => handleChangeRange({ direction: 1 })}
-        onTouchStart={() => Haptics.impactAsync(Haptics.ImpactFeedbackStyle.Light)}
-      />
-    </HStack>
-  </Box>
-);
-
-const mapStateToProps = (state: RootState) => ({
-  start: state.firefly.start,
-  end: state.firefly.end,
-  range: state.firefly.range,
-  rangeTitle: state.firefly.rangeTitle,
-});
-
-const mapDispatchToProps = (state: Dispatch) => ({
-  handleChangeRange: state.firefly.handleChangeRange,
-});
-
-export default connect(mapStateToProps, mapDispatchToProps)(RangeTitle);
